Fix model_name problem with dynamic_form
This was blowing up when error_messages_for calls model_name in order to
prepare the default message.  The reason it blew up was because the
class is anonymous, however even if it has an appropriate name, we
actually want the model name from the base class.  Getting this requires
a bit of gnarly meta programming to get the base_class into the closure.