Test against all versions of Rails 5 and Ruby
Later versions of Rails require higher ruby versions; early versions
of Ruby can't run recent Rails, and vice versa. This exclusion matrix
will get more complicated until we drop some compatibility.

Note that in order to avoid having to install a specific version of
bundler for each ruby version, we need to constraint the one installed
to a verson compatible with Ruby 2.1; the latest bundler on the 1.x
branch requires Ruby 2.3 or above.

In order to ensure we have the right version of sqlite3 (since a
specific version is required by these versions of active record), we
define it explicitly in the appraisals file. We can also use this to
add minitest for the Rails 3.2 build automatically.